requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Settings Management
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0

# Image processing
Pillow==10.2.0

# AWS S3
boto3==1.34.34
